Silence gcc warnings.

+++ src/frame.h
@@ -249,19 +249,19 @@ public:
     operator bool() const { return isValid() && isComplete(); }
 
     uint8_t *data(size_t plane = 0) {
-        if (!m_raw || plane >= (AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
+        if (!m_raw || plane >= size_t(AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
             return nullptr;
         return m_raw->extended_data[plane];
     }
 
     const uint8_t *data(size_t plane = 0) const {
-        if (!m_raw || plane >= (AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
+        if (!m_raw || plane >= size_t(AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
             return nullptr;
         return m_raw->extended_data[plane];;
     }
 
     size_t size(size_t plane) const {
-        if (!m_raw || plane >= (AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
+        if (!m_raw || plane >= size_t(AV_NUM_DATA_POINTERS + m_raw->nb_extended_buf))
             return 0;
         AVBufferRef *buf = plane < AV_NUM_DATA_POINTERS ?
                                m_raw->buf[plane] :
@@ -280,7 +280,7 @@ public:
                 total += m_raw->buf[i]->size;
             }
 
-            for (size_t i = 0; i < m_raw->nb_extended_buf; ++i) {
+            for (size_t i = 0; i < size_t(m_raw->nb_extended_buf); ++i) {
                 total += m_raw->extended_buf[i]->size;
             }
         } else if (m_raw->data[0]) {
@@ -298,7 +298,7 @@ public:
                                                nullptr,
                                                linesizes);
             } else if (m_raw->nb_samples && m_raw->channel_layout) {
-                for (size_t i = 0; i < m_raw->nb_extended_buf + AV_NUM_DATA_POINTERS && m_raw->extended_data[i]; ++i) {
+                for (size_t i = 0; i < size_t(m_raw->nb_extended_buf + AV_NUM_DATA_POINTERS) && m_raw->extended_data[i]; ++i) {
                     // According docs, all planes must have same size
                     total += m_raw->linesize[0];
                 }
